+++ build.gradle
@@ -18,7 +18,7 @@ javafx {
 
 application {
     // FIX THIS to your actual package:
-    mainClass = 'duke.Launcher'   // or 'java.Launcher' temporarily if you haven't renamed packages yet
+    mainClass = 'jarvis.Launcher'   // or 'java.Launcher' temporarily if you haven't renamed packages yet
 }
 
 run {
