--- src/main/java/duke/tasks/TaskList.java
@@ -1,44 +1,0 @@
-package duke.tasks;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class TaskList {
-    private final ArrayList<Task> tasks;
-
-    public TaskList() {
-        this.tasks = new ArrayList<>();
-    }
-
-    public void add(Task t) {
-        tasks.add(t);
-    }
-
-    public int size() {
-        return tasks.size();
-    }
-
-    public Task get(int index) {
-        return tasks.get(index);
-    }
-
-
-    public String toNumberedString() {
-        if (tasks.isEmpty()) {
-            return "no tasks yet";
-        }
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < tasks.size(); i++) {
-            sb.append(i + 1).append(".").append(tasks.get(i)).append("\n");
-        }
-        return sb.toString().trim();
-    }
-
-    public Task remove(int idx) {
-        return tasks.remove(idx);
-    }
-
-    public List<Task> asList() {
-        return tasks;
-    }
-}
