--- src/main/java/duke/ui/DarrenAssistantException.java
@@ -1,7 +1,0 @@
-package duke.ui;
-
-public class DarrenAssistantException extends Exception{
-    public DarrenAssistantException(String message) {
-        super(message);
-    }
-}
