+++ src/main/java/jarvis/ui/DarrenAssistantException.java
@@ -1,0 +1,7 @@
+package jarvis.ui;
+
+public class DarrenAssistantException extends Exception{
+    public DarrenAssistantException(String message) {
+        super(message);
+    }
+}
