+++ src/test/java/darren/assistant/tasks/ToDoTaskTest.java
@@ -1,7 +1,7 @@
 package darren.assistant.tasks;
 
-import duke.tasks.Task;
-import duke.tasks.ToDoTask;
+import jarvis.tasks.Task;
+import jarvis.tasks.ToDoTask;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
